+++ back-end/src/main/java/com/ring/bookstore/dtos/ReviewDTO.java
@@ -3,13 +3,13 @@ package com.ring.bookstore.dtos;
 import java.time.LocalDateTime;
 
 public record ReviewDTO(
-        Integer id,
+        Long id,
         String content,
         Integer rating,
         LocalDateTime date,
         String userName,
-        Integer userId,
+        Long userId,
         String bookTitle,
-        Integer bookId) {
+        Long bookId) {
 
 }
