+++ back-end/src/main/java/com/ring/bookstore/repository/OrderReceiptRepository.java
@@ -12,28 +12,28 @@ import org.springframework.stereotype.Repository;
 import com.ring.bookstore.model.OrderReceipt;
 
 @Repository
-public interface OrderReceiptRepository extends JpaRepository<OrderReceipt, Integer>{
+public interface OrderReceiptRepository extends JpaRepository<OrderReceipt, Long>{
 	
 	@Query("""
 	select case when count(or) > 0 then true else false end 
 	from OrderReceipt or join OrderDetail od on or.id = od.order.id
 	where or.user.userName = :userName and od.book.id = :id
 	""")
-	boolean existsByUserBuyBook(Integer id, String userName); //Check if user have bought this book before
+	boolean existsByUserBuyBook(Long id, String userName); //Check if user have bought this book before
 	
-	Page<OrderReceipt> findAllByUser_Id(Integer id, Pageable pageable); //Get orders from user's {id}
+	Page<OrderReceipt> findAllByUser_Id(Long id, Pageable pageable); //Get orders from user's {id}
 	
 	@Query("""
 	select o from OrderReceipt o join OrderDetail od on o.id = od.order.id 
 	where od.book.seller.id = :id
 	""")
-	Page<OrderReceipt> findAllBySeller(Integer id, Pageable pageable); //Get orders by seller's {id}
+	Page<OrderReceipt> findAllBySeller(Long id, Pageable pageable); //Get orders by seller's {id}
 	
 	@Query("""
 	select o from OrderReceipt o join OrderDetail od on o.id = od.order.id 
 	where od.book.id = :id
 	""")
-	Page<OrderReceipt> findAllByBookId(Integer id, Pageable pageable); //Get orders with book's {id}
+	Page<OrderReceipt> findAllByBookId(Long id, Pageable pageable); //Get orders with book's {id}
 	
 	@Query("""
 	select month(o.oDate) as name, coalesce(sum(o2.amount) , 0) as books, coalesce(sum(distinct o.total) , 0) as sales 
@@ -48,5 +48,5 @@ public interface OrderReceiptRepository extends JpaRepository<OrderReceipt, Inte
 	where b.seller.id = :id
 	group by month(o.oDate)
 	""")
-	List<Map<String,Object>> getMonthlySaleBySeller(Integer id); //Get monthly sale by seller's {id}
+	List<Map<String,Object>> getMonthlySaleBySeller(Long id); //Get monthly sale by seller's {id}
 }
